feat(data): refactor NocoDB related function into its own class
